requirements_updated.txt
Django>=4.0
djangorestframework
psycopg2-binary
djangorestframework_simplejwt
